+++ services/structuralvariation/filt3r/app/bin/filt3r.smk
@@ -481,7 +481,7 @@ onsuccess:
 		print('[INFO] Skipping file copy due to NOCOPY option')
 
 	# Optionally, perform DEPOT_DIR copy
-	if config['DEPOT_DIR'] and outputDir != depotDir:
+	if config['DEPOT_DIR'] and outputDir != depotDir and not config['NOCOPY']:
 		for sample in sample_list:
 			shell(f"rm -f {depotDir}/{sample}/{serviceName}/* || true")
 		shell("rsync -azvh --include={include} --exclude='*' {resultDir}/ {depotDir}")
